requirements.txt
asgiref==3.3.1
Django==3.1.7
django-autoslug==1.9.8
django-crontab==0.7.1
django-model-utils==4.1.1
django-timestampedmodel==0.1.0
djangorestframework==3.12.2
psycopg2==2.8.6
python-decouple==3.4
pytz==2021.1
selenium==3.141.0
sqlparse==0.4.1
urllib3==1.26.4
